be/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
Django==3.0.6
django-cors-headers==3.2.1
djangorestframework==3.11.0
Flask==1.1.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
optionaldict==0.1.1
Pillow==7.1.2
pycparser==2.20
PyMySQL==0.9.3
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
wechatpy==1.8.13
Werkzeug==1.0.1
xmltodict==0.12.0
mysqlclient==1.4.6
dicttoxml==1.7.4
django-apscheduler==0.2.13